feat #0008 Places api 연동 성공
